Move EnsureProxyPermissions to MiscUtils
