+++ PluralKit.Bot/Utils/MiscUtils.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Net.Sockets;
 using System.Threading.Tasks;
 
+using Discord;
 using Discord.Net;
 
 using PluralKit.Core;
@@ -27,6 +28,30 @@ namespace PluralKit.Bot
             if (e is TaskCanceledException) return false;
             
             // This may expanded at some point.
+            return true;
+        }
+
+        public static async Task<bool> EnsureProxyPermissions(ITextChannel channel)
+        {
+            var guildUser = await channel.Guild.GetCurrentUserAsync();
+            var permissions = guildUser.GetPermissions(channel);
+
+            // If we can't send messages at all, just bail immediately.
+            // TODO: can you have ManageMessages and *not* SendMessages? What happens then?
+            if (!permissions.SendMessages && !permissions.ManageMessages) return false;
+
+            if (!permissions.ManageWebhooks)
+            {
+                throw Errors.MissingPermissions("Manage Webhooks", "proxy messages");
+                return false;
+            }
+
+            if (!permissions.ManageMessages)
+            {
+                throw Errors.MissingPermissions("Manage Messages", "delete the original trigger message");
+                return false;
+            }
+
             return true;
         }
     }
